+++ src/helpers/style-dictionary/filters.js
@@ -1,0 +1,38 @@
+/**
+ * used in build.js
+ * filters will be applied before formatting
+ *
+ * order of operations: filters > transforms > formats
+ *
+ * https://amzn.github.io/style-dictionary/#/formats?id=filtering-tokens
+ */
+
+const StyleDictionary = require("style-dictionary");
+
+// filter borders and borderWith
+StyleDictionary.registerFilter({
+  name: "custom/filter/borders",
+  matcher: (token) => {
+    return token.type === "borderRadius" || token.type === "borderWidth";
+  },
+});
+
+// filter typography
+StyleDictionary.registerFilter({
+  name: "custom/filter/typography",
+  matcher: function (token) {
+    return token.attributes.category === "typography";
+  },
+});
+
+// filter boxShadows, colors, and opacity
+StyleDictionary.registerFilter({
+  name: "custom/filter/themeTokens",
+  matcher: (token) => {
+    return (
+      token.attributes.category === "boxShadows" ||
+      token.attributes.category === "colors" ||
+      token.attributes.category === "opacity"
+    );
+  },
+});
